[batch] Enforce abstract properties on driver subclasses (#13186)
Abstract properties need to be declared [this
way](https://stackoverflow.com/questions/2736255/abstract-attributes-in-python)
to be enforceable. If I delete `self.billing_manager` in one of the
driver subclasses it still typechecks but would error at runtime.

I also deleted the `create` method on the `CloudDriver` interface
because it is never really used as an interface method (we only ever
explicitly use the method from the concrete classes). As such it doesn't
really enforce a contract and can be a little restrictive (in terra
there is no sensible `credentials_file` input to such a method to create
a terra driver.