[infra] Add test-dev GSA to gcp-broad (#13197)
Currently hail-vdc uses the GSA for the test user as the GSA for the
test-dev user. We should be using different identities. This is done
correctly in the other terraform so the following steps only apply to
`hail-vdc`. After this change is applied and a new identity for test-dev
is created, I will:

- Create a key for the new service account and update `test-dev-gsa-key`
with that key
- Update the auth database to set the `hail_identity` for test-dev to
the new identity