[batch] Add new list_jobs query interface to REST calls (#12748)
This PR introduces Search v2.0 for querying jobs in a batch with the
REST interface. The UI additions will come in a separate PR as well as
the upgraded search for querying batches. We can decide whether to
upgrade CI to v2. I think for backwards compatibility we need to have
the default remain "1" in perpetuity in the batch client and you have to
explicitly say you want version 2 as an argument if you want the new
query language.

If we want to add tests for the SQL, I can do that once what I have
currently is passing and we're happy with the code.