[batch] JVM container logs streamed with fluentd (#13190)
The goal of this PR is to have all of the JVM container logs available
where all the worker logs are. I tagged the entries with "worker.log" so
they show up with the other worker log entries. However, it's plain text
with no timestamp. We can improve the formatting as a separate project.
Notice the two entries with "*" on the left instead of the normal "I".

The design choice I made is to have the JVM containers write to a
location that is static. We cannot easily change the fluentd
configuration dynamically. It requires restarting the daemon which takes
1.5 seconds. Furthermore, the configuration for fluentd is on /etc/ on
the host which the batch worker container cannot access. Hence, why I
took the approach of specifying it in the startup script at known
locations.

Before we merge this, I'd like to confirm that (a) we want these logs
and (b) they don't contain any secrets.
<img width="1585" alt="Screenshot 2023-06-16 at 4 06 43 PM"
src="https://github.com/hail-is/hail/assets/1693348/0ce9f7dc-1188-4c66-ae6f-83fcc3744f95">