[batch] Mount worker deploy config instead of using k8s secret (#13203)
Once the current auth overhaul is through and we go "keyless", the only
secret left that we mount into user jobs is the deploy config, which at
that point feels kind of silly. It's also nearly always the same deploy
config value that we serialize and write to a file for every job. It
seems cleaner and simpler to me that we create one deploy config for the
worker, and the worker readonly mounts that config into every job.

Note that this is overridable so that any pre-existing jobs that specify
a deploy-config secret and all the special deploy-config stuff that we
do in build.yaml should not be affected