[query] refactor BaseIR so children only exposes iterable interface (#13214)
This is a simple tidying refactoring to hide the `children` array from
consumers of `BaseIR`. My main motivation is to make it easier to
explore other ir data structure designs, and to migrate to a new design
in the future, e.g. to allow for in-place mutation without requiring
large-scale changes to every compiler pass, and to simplify how we
encode binding structure.