[auth] Add internal hail identity UIDs to the users table (#13207)
In order to start using Google or AAD access tokens instead of
hail-minted tokens, we need to be able to identify a service account in
the system by its UID at their identity provider. In Google, this UID is
the `uniqueId` field of the Service Account. In AAD, it is the Service
Principal Object ID. In an upcoming change, we'll update the user
creation process to add this ID upon user creation, at which point we
will be able to safely remove this code that updates existing records.

I've marked this PR as `full-deploy` so the AUS folks can roll this
commit out specifically before this column is relied on. This way we can
safely remove the loop in a follow-up PR and know they will have run
this code to populate the column.