[batch] Add identifying metadata to Batch requests (#13219)
I realize this looks like a lot of code changes, but it's mostly copying
and pasting two SQL procedures and changing one line in each.

This adds 4 bits of metadata to requests that then can be queried as
extra metadata:
- batch_id
- job_id
- batch_operation
- job_queue_time

Should be self-explanatory except job_queue time is the time in which
the job is first set to ready to when it was scheduled on the worker
(exact moment is when the job config is made to send to the worker).

Example logging query. Note that the search on "batch_id" is not
optimized so you definitely want to add some kind of time limit that's
short on the window to search. I can add my Python script that scrapes
these logs and makes a Plotly figure in a separate PR once this goes in.

```
(
resource.labels.container_name="batch"
resource.labels.namespace_name="{namespace}"
) OR (
resource.labels.container_name="batch-driver"
resource.labels.namespace_name="{namespace}"
) OR (
resource.type="gce_instance"
logName:"worker.log"
labels."compute.googleapis.com/resource_name":"{namespace}"
)
jsonPayload.batch_id="{batch_id}"
timestamp >= "{start_timestamp}" {end_timestamp}
```