[ci] Add test identities for services in test namespaces (#13202)
Identities in test namespaces cannot share the same underlying cloud
identity if we want to identify requests with cloud access tokens. This
also means the `test` account does not need to have the union of roles
of the other robot accounts, but pruning of the `test` account's roles
is left until after this PR merges so we can properly assess which roles
are still in use by the `test` account.