[batch] Add Azure debugging information when deleting deployments (#13231)
I need this extra debugging information to understand what is going on
in Azure with deleted VMs still showing up in the portal with
ResourceNotFound errors. Miah and Greg are running into this same
problem in their deployment. My guess is what is happening is the worker
is active and working fine, but then the deployment gets "Canceled"
because the OMSAgent takes too long to deploy. So our loop then cancels
the deployment which messes up the state in Azure of the already
deployed and running VM.

I popped the parameters from the deployment result in case it contains
sensitive data (I'm mainly worried about any private SSH keys).