[batch] Fix list batches query and test (#13237)
The test that lists batches timed out. The main problem is the limit in
the aioclient used by the test_batch tests was passing a string rather
than an integer. I assumed downstream the function was passing an
integer. Therefore, we were doing this:

batch_id < "137"
and not batch_id < 137.

So the query was running forever and scanning all batches from the test
user.

I also was missing a tag annotation on the queries, but that was not
causing the timeout.