[batch] Increase max worker idle time to 2 mins in test and dev (#13314)
I observed a cluster with it set to the default idle time of 30 seconds
in Azure and the workers were continuously thrashing leading up to 49
instances being created over the course of a PR. With an idle time of
120 seconds, there was no thrashing and 28 instances were created over
the course of the PR (16 standard + job private etc.). The cluster
nicely scaled down at the end of the PR. It looked like a couple of
times the `standard-np` pool scaled up and then scaled down so I assume
the `standard` pool wasn't at full capacity while that was happening. It
might be worth configuring the `standard-np` pool to be 4 or 5 standing
instances with 16 cores and see what happens -- that might help as well.