[security] delete after 5 days, ignore system managed, consider two special secrets
A job running for five days *could* cause an issue here, so the operator should still be cognizant of that. The
move to access tokens will eliminate that concern anyway.

We should not consider system-managed keys because those are managed by Google and not even visible in the
UI. We are not obligated to rotate them (nor are we able, afaik).

There are two special secrets which do not conform to the `key.json` naming scheme for their
secret files. I added them as special cases.