[query] removes unnecessary serialization in is.hail.rvd.RVD.localSort (#13313)
On the current release (0.2.119), `is.hail.rvd.RVD.localSort` attempts
to serialize an entire `is.hail.rvd.RVD` object when called, which
causes some operations, like reading a set of VCF files, to fail. That
particular operation no longer fails due to some combination of the
changes made to reading VCF files in
https://github.com/hail-is/hail/pull/13206 and
https://github.com/hail-is/hail/pull/13229, but the serialization could
still cause issues if `is.hail.rvd.RVD.localSort` is called in another
context.

See
https://hail.zulipchat.com/#narrow/stream/123011-Hail-Query-Dev/topic/Possible.20bug.20in.20Hail.200.2E2.2E119/near/378610151
for more information.