fix rootdir for doctest and also restore use of SKIP_OUTPUT_CHECK
The essential difference is the rootdir. If you specify a path of test files which shares with your
current working directory at least one common directory other than `/`, the rootdir will be the
common directory. For example:

    cd /foo
    pytest /foo/bar/baz

The rootdir will be set to `/foo`. In contrast:

    cd /
    pytest /foo/bar/baz

or

    cd /applepie
    pytest /foo/bar/baz

the rootdir will be set to /foo/bar/baz.

Pytest, apparently, will load `site-packages/hail/conftest.py` if the rootdir is a parent of
`site-packages/hail`. In the latter two cases, this is not the case. Consider the real example:

    cd /
    pytest /Users/dking/miniconda3/lib/python3.10/site-packages/hail/docs

the rootdir is

    /Users/dking/miniconda3/lib/python3.10/site-packages/hail/docs

which is not an ancestor of the file

    /Users/dking/miniconda3/lib/python3.10/site-packages/hail/conftest.py

Interestingly, pytest says the following about where to put conftest.py. I'm not sure why they think
it is a bad idea to have per-directory conftest.py. It seems to me like just a choice of whether you
want global or per-directory plugins/fixtures.

https://docs.pytest.org/en/latest/how-to/writing_plugins.html#conftest-py-plugins:

> Note that pytest does not find conftest.py files in deeper nested sub directories at tool
> startup. It is usually a good idea to keep your conftest.py file in the top level test or project
> root directory.

Immediately after this section they discuss [local per-directory
plugins](https://docs.pytest.org/en/latest/how-to/writing_plugins.html#conftest-py-local-per-directory-plugins)

There is also a different page with specific information about when conftest.py is loaded:

https://docs.pytest.org/en/stable/how-to/doctest.html#doctest-namespace-fixture

> Note that like the normal conftest.py, the fixtures are discovered in the directory tree conftest
> is in. Meaning that if you put your doctest with your source code, the relevant conftest.py needs
> to be in the same directory tree. Fixtures will not be discovered in a sibling directory tree!

Here is some extensive discussion of how the rootdir is discovered:

https://docs.pytest.org/en/stable/reference/customize.html#finding-the-rootdir

Here is the algorithm which finds the rootdir from args:

> If -c is passed in the command-line, use that as configuration file, and its directory as rootdir.
>
> - Determine the common ancestor directory for the specified args that are recognised as paths that exist in the file system. If no such paths are found, the common ancestor directory is set to the current working directory.
>
> - Look for pytest.ini, pyproject.toml, tox.ini, and setup.cfg files in the ancestor directory and upwards. If one is matched, it becomes the configfile and its directory becomes the rootdir.
>
> - If no configuration file was found, look for setup.py upwards from the common ancestor directory to determine the rootdir.
>
> - If no setup.py was found, look for pytest.ini, pyproject.toml, tox.ini, and setup.cfg in each of the specified args and upwards. If one is matched, it becomes the configfile and its directory becomes the rootdir.
>
> - If no configfile was found and no configuration argument is passed, use the already determined common ancestor as root directory. This allows the use of pytest in structures that are not part of a package and don’t have any particular configuration file.

In local development, you and I probably both use editable installs. In that case, the setup.cfg
file in `hail/python` makes `hail/python` the rootdir. That includes `hail/python/hail/conftest.py`
so it gets loaded.

The fix seems to be to explicitly specify the rootdir.