[azure] Add connection timeouts for Azure FS in both Python and Scala (#13344)
I hope this helps. The regular `timeout` parameter on the Python methods
like `stage_block` is for the server-side timeout of the operation. We
want the client side timeout to be shorter. I picked 5 seconds, but that
was arbitrary. I'm not sure I implemented the Scala code correctly, but
that's along the right lines of what needs to happen.