[qob] disable QoB PR tests (keep deploy and dev) (#13353)
For added context see: https://github.com/hail-is/hail/issues/13351

cc: @patrick-schultz @chrisvittal @daniel-goldstein @ehigham
@iris-garden

I actually don't think we need to notify anyone because we'll still get
errors if something goes wrong in default. If main starts failing, we
have the slightly annoying situation of needing to run tests in a
one-off manner to verify and we might need to block a release until we
can revert.