[fs] do not unlink the directory if there was an exception (#13362)
I created this bug when I added exception handling. I incorrectly put
the exception handling *after* the `rmdir`. If an exception occurred
while removing one of the children of a directory, in all likelihood,
the directory is non-empty.

With this change, we raise any exceptions before unlinking the
directory. It is the responsibility of the caller to decide what to do
if we were unable to remove one of our children.

I saw this while debugging another transient error on another PR:
https://github.com/hail-is/hail/issues/13361.