[auth] Dont use the auth client for authz in the auth service (#13330)
Came across this while doing #13307. The `AuthClient` decorators are
used by services like Batch and CI to protect endpoints behind an
authentication/authorization check. Those decorators might make a
network request to the auth service to see who the user is and whether
they're a developer. I realized that these decorators are also used to
protect some endpoints on the auth service too, which albeit uniform
felt silly to me (make a network request back to yourself instead of
just making a function call). Creating new decorators to use within the
auth service is a bit more code, but I personally think it's a much
easier code path to follow.


Only the last commit is this PR, as it's stacked.