[batch] Specify batch-driver as the default container in the batch-driver pod (#13374)
`kubectl` commands like `exec` that deal with containers in pods will
generally default to the first container in the pod. The `batch-driver`
deployment has two containers, the envoy sidecar and the batch-driver
python app. Since the `envoy` container is listed first, `kssh` and
`sync.py` default to that container, but we basically always intend
`batch-driver`. Instead of switching up the order, or adding a bunch of
explicit container specification, this annotates `batch-driver` as the
default container in the pod, so `kssh` and `sync.py` work with the
batch driver.