[azure] Optimize file status operations (#13368)
I think this change will help the number of operations we're making
substantially. My Scala skills are not great, so I don't know if this is
written correctly.

Basically, we were making a call to list the blobs recursively to test
if the path was a directory which was streaming through the first 5000
records. I made the page size equal to 1 record as we don't care about
all records. The next thing I did was to just get the blob properties
rather than calling exists + get blob properties. So that will cut the
number of HTTP calls by half for every blob. Lastly, listing items in a
directory which is used for globbing was making a call to get the
metadata for each file and then it was making the 3 API calls above to
check whether it's a directory, whether it exists, and what the blob
properties are. All of this information is in the original result from
listing the blobs in the hierarchy so I just use that information
directly.