[batch] mitigate bug in siwei's pipeline
Before open batches, the `n_jobs` of a batch was a constant known before any jobs were added. Moreover,
we did not start scheduling jobs until all the jobs were added to the database. Therefore,
it was always safe to assume that the final "bunch" of jobs in the database was the last
"bunch" ergo it spanned from its `start_job_id` to the job with id `n_jobs` (nb: job ids are
1-indexed).

When open batches were added, the `n_jobs` became a mutable value. Moreover, `n_jobs` includes
jobs in bunches *which have not yet been added to the database*. In particular, suppose two clients
are each submitting a bunch of size 10. Each client independently "reserves" 10 job
slots by atomically incrementing `n_jobs` by ten. `n_jobs` is now 20. Further suppose that
the first bunch is added to the database and begins scheduling before the second bunch
is added to the database. In this case, when calculating the size of this bunch (for use
in the bunch cache, and *only* in the bunch cache), we see that this is the last (and
only) bunch in the database and assume that `n_jobs` is the last job id in this bunch.
This is incorrect because `n_jobs` includes the not-yet-visible second bunch.