[compiler] fix TableUnion lowering (#13414)
fixes #13407

CHANGELOG: Resolves #13407 in which uses of `union_rows` could reduce
parallelism to one partition resulting in severely degraded performance.

TableUnion was always collapsing to a single partition when the key was
empty. This adds a special case handling, which just concatenates
partitions.

The body of the resulting TableStage is a little hacky: it does a
StreamMultiMerge, but where exactly one input stream is non-empty. I
think that should have fine performance, and I didn’t see any simpler
ways to do it.