[query] Avoid Class A Operations.
CHANGELOG: Hail Query-on-Batch previously used Class A Operations for all interaction with blobs. This change ensures that QoB only uses Class A Operations when necessary.

Inspired by @jigold 's file system improvement campaign, I fell down the rabbit hole of not issuing
"list" operations unless we really need to know if a file is a directory. This should partly help
reduce the flakiness in Azure (which is tracked in #13351) as well as high costs in Azure both of
which are at least partly suspected to be caused by our frequent use of "list" operations.