[query] force `aggregate_cols` to be local (#13405)
CHANGELOG: MatrixTable.aggregate_cols no longer forces a distributed
computation. This should be what you want in the majority of cases. In
case you know the aggregation is very slow and should be parallelized,
use mt.cols().aggregate instead.

Most of the time, `aggregate_cols` will be much faster performing the
aggregation locally. Currently, we generate a `TableAggregate` over a
`TableParallelize` of the columns. We shouldn't try to optimize that to
a local computation during compilation; `TableParallelize` should
express the intent that the computation is expensive and really should
be parallelized. This should be considered part of the semantics the
compiler must preserve.

This PR changes `aggregate_cols` to explicitly generate a local
computation using `StreamAgg` (which was only exposed in Python
relatively recently, which is why we haven't made this change sooner).
Longer term, aggregating columns should probably get its own IR node,
especially once we start partitioning along columns.