[hailctl] Dont change the current python environment in test fixture (#13457)
Sorry missed this in the initial PR. `os.environ` is global to the whole
python session so setting it directly could inadvertently affect the
environment for unrelated tests. There seems to be safe ways to do this
using `monkeypatch` and whatever `CliRunner` is doing, but I decided
just not to set it.