[hailctl] clarify suggested use of -- and allow it in hailctl batch submit (#13447)
CHANGELOG: Resolves #13446. In all three submit commands (batch,
dataproc, and hdinsight), Hail now allows and encourages the use of --
to separate arguments meant for the user script from those meant for
hailctl. In hailctl batch submit, option-like arguments, for example
"--foo", are now supported before "--" if and only if they do not
conflict with a hailctl option.