[batch_client.aioclient] Get rid of BatchBuilder (#13458)
This PR refactors the aioclient to merge the functionality of Batch and
BatchBuilder into just a single Batch object. The reason for making this
change is to make adding job groups simpler. I will follow up with a
change to Jobs after this merges. I apologize for the number of line
changes. Most are just renaming `bb -> b` in the tests.