[query] improve type hinting for tables
Without these changes, my IDE cannot deduce that the return value of, say, `select`
is a `Table` which prevents jump-to-definition on subsequent method calls.

I also received an error about `Table.collect` returning an `Expression` which
it will only do when `_localize=False`. `overload` teaches this fact to the
type system.