[fs] remove Azure _debug parameters
This was added in https://github.com/hail-is/hail/pull/12421 to help debug errors in `testSeekMoreThanMaxInt`.
We have not seen that transient error in a while and I think the switch to https://github.com/hail-is/hail/pull/12590
might have fixed some underlying misuse of `AppendBlobClient` (by not using it).