[ci/ui] fixes 500 when there are no jobs to display (#13254)
This change sorts the jobs for a PR correctly in the UI, so that all
jobs are placed under a header with the correct state (see screenshot).
Because jobs are split into separate tables by their state,
`focusOnSlash` has been removed from the relevant CI pages, since it is
unknown which tables will exist on the page until it is rendered. It
also fixes the error that was causing the server to return a 500 when
there were no jobs yet (e.g. when a retry had just been requested),
which was caused by an assumption in the original job filtering code
that there would always be at least one job to display. It also passes
the list of developers through to the underlying `PR` object that a
`WatchedBranch` has, which was missed in
https://github.com/hail-is/hail/pull/13398 but is required for CI to
display the PR page correctly in dev deploys only.

<img width="714" alt="Screenshot 2023-08-22 at 15 08 48"
src="https://github.com/hail-is/hail/assets/84595986/6bcab38c-a48a-4e10-8f69-770d35ea51b7">