[dev-docs] Add document detailing git practices (#13492)
I feel like most of us do most of this but we don't have anything to
point new developers to. I'm also kind of curious if this is what other
folks' understanding of a best practice. Maybe something we can
~~bikeshed~~ bring up in meeting?

I also reorganized a bit while I was in the `dev-docs` directory.