[query] improve type hinting for tables (#13468)
Without these changes, my IDE cannot deduce that the return value of,
say, `select` is a `Table` which prevents jump-to-definition on
subsequent method calls.

I also received an error about `Table.collect` returning an
`ArrayExpression` which it will only do when `_localize=False`.
`overload` teaches this fact to the type system.

I didn't use the specific expression in `base_expression.py` because
that would require an import circularity.