[batch] Drop unused SQL functions (#13510)
I want to cleanup some of the old database functions as they're no
longer used (`close_batch`) and not tested / not used in production /
hard to maintain (`recompute_incremental`). The `recompute_incremental`
was there in the early days of Batch when we needed to recompute the
user_inst_coll_resources table etc. for a single migration to improve
cancelling. We have not used it since.

I double checked `close_batch` isn't used in the Python code, but would
appreciate it if you could also check as well.