[auth] Fix the auth rest API for developers (#13515)
I must have broken this during a recent refactor of the decorators. When
making requests to an internal namespace, the default namespace's auth
token is set in the `X-Hail-Internal-Authorization` header and the dev
namespace's token is in the `Authorization` header. The dev namespace
needs to know *not* to pick up the `X-Hail-Internal-Authorization`
header.