[lint] Swap mypy for pyright (#13437)
`pyright` is an alternative to `mypy` for type checking python. It's
also a pretty solid language server. While they both adhere to the PEPs
for type checking, they [have a few
differences](https://github.com/microsoft/pyright/blob/main/docs/mypy-comparison.md#differences-between-pyright-and-mypy).
That doc is worth a read, but the parts that I found most compelling are
pyright's return-type inference that mypy doesn't do and that pyright
differentiates between an `Unknown` type and an `Any` type. `mypy`, for
the most part, treats variables as `Any` if it can't figure out what
type they are, and as a result lets a lot of behaviors on `Any`
variables slide. `pyright` tends to be much stricter (even when not on
strict mode!) and while some of the assertions that I had to make here
to appease `pyright` are a bit noisy, it's also pointed out a lot of
areas where we can be stricter with our typing and then clean up code
because some assertions are caught instead by the type system. It also
did a good job informing of bad import practices like how `hl.utils` was
not actually exported by the `hail` module but incidentally imported by
some file in the repo.

---------

Co-authored-by: Dan King <dking@broadinstitute.org>