[batch] More debugging information for network namespace errors (#13509)
In a previous PR we added exception logging when tasks on a
`BackgroundTaskManager` fail. I'm not seeing any of those messages in
the worker logs, suggesting that for some reason the Tasks to free a
network namespace when a job is finished are not getting created in the
first place. I don't see how in the code we could somehow *not* free the
network namespace, but I'm hoping that some of these diagnostics shed
some light. Also open to suggestions for where else might be a good spot
to add more logging.